Magic to help sync STDERR/STDOUT/WARNINGS
This may be too magical. I might make it an option, one that is opt-in
instead of opt-out.

This makes it so that STDERR/STDOUT and warnings are first class events
that nest properly in subtests and all that good stuff.

Problem is this easily breaks any tests that do anything with
STDERR/STDOUT. Even some of Test2::Harness's tests needed to add some
strategic untie() calls to make it work.

Before merging I will probably make --tie-io and --no-tie-io options for
this, and probably also some harness directives as well so I can clear
those untie() calls.